requirements.txt
numpy==1.19.1
Pillow==7.1.2
scipy==1.5.0
imagecodecs==2020.5.30
tqdm==4.46.1
tensorflow-gpu==1.14.0
Keras==2.3.1
h5py==2.10.0
csbdeep==0.6.0
bson==0.5.10
scikit-image==0.17.2
scikit-learn==0.24.2
pandas==1.1.5
tifffile==2020.9.3
pandas==1.1.5